berry-ai/backend/requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
python-multipart==0.0.20
Pillow==11.1.0
numpy==2.2.3
ultralytics==8.3.82
opencv-python-headless
